LayoutTests/http/tests/events/device-orientation-motion-enumerable-expected.txt

Test that ondevicemotion and ondeviceorientation are enumerable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.getOwnPropertyNames(window).includes('ondeviceorientation') is true
PASS Object.getOwnPropertyNames(window).includes('ondevicemotion') is true
PASS successfullyParsed is true

TEST COMPLETE
